Added user authentication and authorization. Implemented sessions and validation for registration and login.

+++ src/controllers/notesController.js
@@ -2,27 +2,27 @@ import {Note} from "../models/note.js"
 import createHttpError from "http-errors";
 
 export const getAllNotes = async (req, res) => {
-  const { page = 1, perPage = 10, tag, search } = req.query;
+  const {page = 1, perPage = 10, tag, search} = req.query;
   const skip = (page - 1) * perPage;
 
-  const notesQuery = Note.find();
+  const notesQuery = Note.find({userIs: req.user._id});
 
   // Будуємо фільтр
-  if(tag) {
+  if (tag) {
     notesQuery.where("tag").equals(tag);
   }
 
-  if(search) {
+  if (search) {
     notesQuery.where({
       $or: [
-        { title: { $regex: search, $options: "i" } },
-        { content: { $regex: search, $options: "i" } },
+        {title: {$regex: search, $options: "i"}},
+        {content: {$regex: search, $options: "i"}},
       ],
     })
   }
 
   // Виконуємо одразу два запити паралельно
-  const [ totalNotes, notes ] = await Promise.all([
+  const [totalNotes, notes] = await Promise.all([
     notesQuery.clone().countDocuments(),
     notesQuery.skip(skip).limit(perPage),
   ]);
@@ -42,7 +42,7 @@ export const getAllNotes = async (req, res) => {
 
 export const getNoteById = async (req, res) => {
   const {noteId} = req.params;
-  const note = await Note.findById(noteId);
+  const note = await Note.findOne({_id: noteId, userId: req.user._id});
 
   if (!note) {
     throw createHttpError(404, "Note not found");
@@ -53,14 +53,14 @@ export const getNoteById = async (req, res) => {
 
 
 export const createNote = async (req, res) => {
-  const note = await Note.create(req.body);
+  const note = await Note.create({...req.query, userId: req.user._id});
   res.status(201).json(note);
 };
 
 
 export const deleteNote = async (req, res) => {
   const {noteId} = req.params;
-  const note = await Note.findOneAndDelete({_id: noteId});
+  const note = await Note.findOneAndDelete({_id: noteId, userId: req.user._id });
 
   if (!note) {
     throw createHttpError(404, "Note not found");
@@ -73,7 +73,10 @@ export const deleteNote = async (req, res) => {
 export const updateNote = async (req, res) => {
   const {noteId} = req.params;
   const note = await Note.findOneAndUpdate(
-    {_id: noteId}, // search by id
+    {
+      _id: noteId,
+      userId: req.user._id
+    }, // search by id
     req.body,
     {
       returnDocument: "after",
